docs: clarify commit message header
Adds extra details regarding the commit message header.  Updates scopes
to the latest, explicitly marks the scope as optional, expands on when
one would use the generic scopes, and adds a note about how to think
about the commit header.  This is for ease of reading as well as syncing
the scopes with the repo's settings.