feat: overhaul item capture popup
Previously, the item capture grid was the same grid as the items only
tracker.  This caused the item capture popup to show things that
shouldn't/couldn't be captured (ex. Star Spirits).  A new component was
created to separate elements that shouldn't be in the capture popup.

It also showed the Warehouse Key as a capturable item, however, players
wouldn't be able to know exactly which key the generic key sprite is
hiding (Worthless Prison Key or valuable Stockroom/Warehouse Keys?).
This replaces the WH-overlaid key with a generic key sprite.