refactor(ch5): remove move raven statue flag
This is the start of an effort to remove tracker flags that are
unnecessary, like this flag. This flag tells the tracker the player
moved the raven statue in Chapter 6 with the Jade Raven.

This flag adds no additional logic to the tracker other than adding
human toil by requiring an extraneous click. It also disguises a number
of checks behind what looks like one, making it seem that the player may
only have one accessible check in an area when there's actually more.