refactor(logic): simplify always active partners
Instead of having a separate line in the access rules for the always
active partners setting flag making the rules harder to read.  This
commit DRYs out the access rules by using the new partner() script.

The new partner() script checks for the always active partners setting
and immediately returns true if it's active, otherwise queries the
tracker to see if the player has collected the required partner.