refactor(ch4): simplify logic for color puzzle
In an effort to remove extra flags that don't provide any additional
value to the tracker, the flag to mark the toybox color puzzle as
solvable is removed in this commit.  The randomizer expects the player
to have both the Dictionary and the Mystery Note _turned-in_ to Russ T.
to progress past the color puzzle.

Because the solution to the puzzle is not currently randomized, the
tracker accepts having only a hammer as a "sequence break" option; this
logic can be kept even if the solution is randomized in the future, as a
player could still brute force it (albeit slowly).

Assistance with vanilla progression of the main game is not in the scope
of the tracker, which can lead to bloat and confusion.  Players should
reference online walkthroughs for help with the vanilla story.