refactor(ch2): remove kooper bridge flag
This is another extraneous flag that doesn't add any value to the
tracker. The randomizer only expects the player to have Kooper, who is
needed to hit the switch on the other side of the river.

Assistance with vanilla progression of the main game is not in the scope
of the tracker, which can lead to bloat and confusion.  Players should
reference online walkthroughs for help with the vanilla story.