refactor(checks): move shops inside parent dungeon
Instead of having a separate Location for item shops (or shop-likes)
that are inside dungeons, this moves shops to be a child of it's parent
dungeon.  Unlike Sections, children can have a separate map location as
they are of type Location.

This ensures that the parent dungeon's access rules are also applied to
the shops where it wasn't previously.