fix: require parakarry in OBK_06,IWA_03,ARN_04,DGB
OBK_06: The library shelves are one Parakarry's width apart.

IWA_03: Coin ring on the left side of the map is past a Parakarry gap.

ARN_04: All checks are past a small gap that requires Parakarry to cross

DGB: Tubba Blubba's Castle is past a small gap in ARN_04 that requires
Parakarry to cross.