refactor(ch2): remove raised ruins flag
The flag to mark the dry dry ruins as raised is removed in this commit.
The randomizer expects the player to have the pulse stone.

Assistance with vanilla progression of the main game is not in the scope
of the tracker, which can lead to bloat and confusion.  Players should
reference online walkthroughs for help with the vanilla story.