refactor: detach weight flag from checks
In order to clean up some of the unnecessary flags in the tracker, this
will detach the `flag_attached_weight` from checks that need the
chandelier weighted in Boo's Mansion.

This flag is yet another example of a flag that hides other checks w/o
providing any additional value to the player.